+++ procountor/__init__.py
@@ -2,6 +2,6 @@
 
 """Top-level package for Procountor."""
 
-__author__ = """Joonas Maliniemi"""
-__email__ = 'joonas@vilkas.fi'
+__author__ = """Joonas Maliniemi, Tero Kotti"""
+__email__ = 'joonas@vilkas.fi, tero@vilkas.fi'
 __version__ = '0.2.1'